Addition Of Remote Input Types
Wget has a standard way of formatting time and file sizes. Seconds or
Bytes are strait integers however other data types are an integer
followed by a single letter denoting the datatype.

Fixing this in classes solves the need to do lots of validation checks
that all do the same thing in multiple places and makes it a lot clearer
what classes expect and accept.

Not 100% sure about the class names, but this can always be updated
later on.